refactor(proto): jina primitive types
