+++ jina/drivers/encode.py
@@ -5,7 +5,7 @@ from typing import Sequence
 
 from . import BaseExecutableDriver
 from .helper import extract_docs
-from ..proto.ndarray.generic import GenericNdArray
+from jina.types.ndarray.generic import GenericNdArray
 
 if False:
     from ..proto import jina_pb2
