+++ jina/drivers/multimodal.py
@@ -8,7 +8,7 @@ import numpy as np
 
 from .encode import BaseEncodeDriver
 from ..proto import jina_pb2
-from ..proto.ndarray.generic import GenericNdArray
+from jina.types.ndarray.generic import GenericNdArray
 
 
 def _extract_doc_content(doc: 'jina_pb2.Document'):
