+++ jina/types/ndarray/dense/__init__.py
@@ -1,4 +1,4 @@
-from ... import jina_pb2
+from jina.proto import jina_pb2
 from .. import BaseNdArray
 
 
@@ -10,4 +10,4 @@ class BaseDenseNdArray(BaseNdArray):
     """
 
     def null_proto(self):
-        return jina_pb2.DenseNdArray()
+        return jina_pb2.DenseNdArrayProto()
