+++ tests/integration/evaluation/rank/test_evaluation.py
@@ -4,7 +4,7 @@ import numpy as np
 
 from jina.flow import Flow
 from jina.proto import jina_pb2
-from jina.proto.ndarray.generic import GenericNdArray
+from jina.types.ndarray.generic import GenericNdArray
 
 
 def test_evaluation(tmpdir):
