+++ tests/integration/incremental_indexing/__init__.py
@@ -4,7 +4,7 @@ import numpy as np
 import pytest
 
 from jina.proto import jina_pb2
-from jina.proto.ndarray.generic import GenericNdArray
+from jina.types.ndarray.generic import GenericNdArray
 
 
 @pytest.fixture(scope='function')
