+++ tests/integration/multimodal/test_multimodal_parallel.py
@@ -5,7 +5,7 @@ import pytest
 
 from jina.flow import Flow
 from jina.proto import jina_pb2
-from jina.proto.ndarray.generic import GenericNdArray
+from jina.types.ndarray.generic import GenericNdArray
 
 NUM_DOCS = 100
 cur_dir = os.path.dirname(os.path.abspath(__file__))
