+++ tests/unit/drivers/test_encoder_driver.py
@@ -5,7 +5,7 @@ import numpy as np
 from jina.drivers.encode import EncodeDriver
 from jina.executors.encoders import BaseEncoder
 from jina.proto import jina_pb2
-from jina.proto.ndarray.generic import GenericNdArray
+from jina.types.ndarray.generic import GenericNdArray
 
 
 class MockEncoder(BaseEncoder):
