+++ tests/unit/drivers/test_multimodal_driver.py
@@ -5,7 +5,7 @@ from jina.drivers.multimodal import MultiModalDriver
 from jina.executors.encoders.multimodal import BaseMultiModalEncoder
 from jina.proto import jina_pb2
 from jina.proto import uid
-from jina.proto.ndarray.generic import GenericNdArray
+from jina.types.ndarray.generic import GenericNdArray
 
 
 @pytest.fixture(scope='function')
