+++ tests/unit/drivers/test_vector_fill_driver.py
@@ -5,7 +5,7 @@ import numpy as np
 from jina.drivers.search import VectorFillDriver
 from jina.executors.indexers import BaseIndexer
 from jina.proto import jina_pb2
-from jina.proto.ndarray.generic import GenericNdArray
+from jina.types.ndarray.generic import GenericNdArray
 
 
 class MockIndexer(BaseIndexer):
