+++ tests/unit/drivers/test_vector_search_driver.py
@@ -6,7 +6,7 @@ import pytest
 from jina.drivers.search import VectorSearchDriver
 from jina.executors.indexers import BaseVectorIndexer
 from jina.proto import jina_pb2
-from jina.proto.ndarray.generic import GenericNdArray
+from jina.types.ndarray.generic import GenericNdArray
 
 
 class MockIndexer(BaseVectorIndexer):
