+++ tests/unit/peapods/test_container.py
@@ -12,7 +12,7 @@ from jina.parser import set_pea_parser, set_ping_parser
 from jina.peapods.container import ContainerPea
 from jina.peapods.pea import BasePea
 from jina.proto import jina_pb2, uid
-from jina.proto.ndarray.generic import GenericNdArray
+from jina.types.ndarray.generic import GenericNdArray
 
 cur_dir = os.path.dirname(os.path.abspath(__file__))
