+++ tests/unit/test_is_read_message.py
@@ -5,7 +5,7 @@ from jina.parser import set_pea_parser
 from jina.peapods.pea import BasePea
 from jina.peapods.zmq import Zmqlet
 from jina.proto import jina_pb2
-from jina.proto.message import ProtoMessage
+from jina.types.message import ProtoMessage
 
 
 class MockBasePeaNotRead(BasePea):
